fix: make capabilities.metadata human readable
